pyproject.toml
[project]
name = "itk-test"
version = "0.1.0"
description = ""
authors = [
    {name = "Anton Novojilov",email = "wowlass63@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pytest-asyncio (>=1.1.0,<2.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "mypy (>=1.17.0,<2.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "fastapi-users (>=14.0.1,<15.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "fastapi-users-db-sqlalchemy (>=7.0.0,<8.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
